--- src/component/README.md
@@ -1,11 +1,0 @@
-# src/component
-
-## index.ts
-
-Examples of [Web Components](https://developer.mozilla.org/en-US/docs/Web/Web_Components). 
-
-Web Components consist of three main technologies:
-- [Custom Elements](https://developer.mozilla.org/en-US/docs/Web/Web_Components/Using_custom_elements)
-- [Shadow DOM](https://developer.mozilla.org/en-US/docs/Web/Web_Components/Using_shadow_DOM)
-- HTML templates, i.e. [template](https://www.w3schools.com/tags/tag_template.asp) and [slot](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/slot) tags
-
