+++ src/template/README.md
@@ -2,6 +2,9 @@
 
 ## index.ts
 
-Examples of [using templates and slots](https://developer.mozilla.org/en-US/docs/Web/Web_Components/Using_templates_and_slots)
+[Web Components](https://developer.mozilla.org/en-US/docs/Web/Web_Components) consist of three main technologies:
+- [Custom Elements](https://developer.mozilla.org/en-US/docs/Web/Web_Components/Using_custom_elements)
+- [Shadow DOM](https://developer.mozilla.org/en-US/docs/Web/Web_Components/Using_shadow_DOM)
+- HTML templates, i.e. [template](https://www.w3schools.com/tags/tag_template.asp) and [slot](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/slot) tags
