+++ src/template/index.ts
@@ -1,5 +1,6 @@
 import { sprintf } from "sprintf-js";
 import { Template, TemplateOptions, TemplateVariable } from "./template";
+import { Component } from "./component";
 
 export namespace index {
     let loadTemplateCounter = 1
@@ -22,8 +23,34 @@ export namespace index {
         ]
         options.variables = variables
         let template = new Template(options)
-        template.load("examples/templates/my-paragraph.html");
+        template.load("examples/templates/_my-paragraph.html");
         loadTemplateCounter++
+    }
+
+    export function shadowDOM() {
+        Component.defineCustomElement("my-paragraph", "my-paragraph-sd")
+    }
+
+    export function namedSlots() {
+        // Creating a template with some slots
+        // https://developer.mozilla.org/en-US/docs/Web/Web_Components/Using_templates_and_slots#creating_a_template_with_some_slots
+        customElements.define('element-details',
+            class extends HTMLElement {
+                constructor() {
+                    super();
+                    const template = 
+                        document.getElementById('element-details-template') as 
+                        HTMLTemplateElement;
+                    let templateContent = template.content;
+                    const shadowRoot = this.attachShadow({ mode: 'open' });
+                    shadowRoot.appendChild(templateContent.cloneNode(true));
+                }
+            }
+        );
+    }
+
+    export function componentUsingVueJS() {
+        console.info("TODO")
     }
 }
