+++ src/template/utils.ts
@@ -1,0 +1,4 @@
+export enum ShadowMode {
+    open = "open",
+    closed = "closed",
+}
