+++ www/examples/templates/index.html
@@ -21,12 +21,12 @@
       <a href="https://www.w3schools.com/js/js_debugging.asp">Dev Console</a>
     </p>
     <p>
-      Also see <a href="/examples/web-components">web-components</a>
+      Also see <a href="/examples/templates/web-components.html">web-components</a>
     </p>
     <ul>
-      <li><a href="javascript:" x-on:click="examples.template.template()">
+      <li><a href="javascript:" x-on:click="app.examples.template.template()">
           Template</a></li>
-      <li><a href="javascript:" x-on:click="examples.template.loadTemplate()">
+      <li><a href="javascript:" x-on:click="app.examples.template.loadTemplate()">
           Load template from file</a></li>
     </ul>
   </div>
