+++ www/examples/templates/web-components.html
@@ -24,9 +24,9 @@
       Also see <a href="/examples/templates">templates</a>
     </p>
     <ul>
-      <li><a href="javascript:" x-on:click="examples.component.shadowDOM()">
+      <li><a href="javascript:" x-on:click="app.examples.template.shadowDOM()">
           Shadow DOM</a></li>
-      <li><a href="javascript:" x-on:click="examples.component.namedSlots()">
+      <li><a href="javascript:" x-on:click="app.examples.template.namedSlots()">
           Named slots</a></li>
     </ul>
   </div>
