+++ www/index.html
@@ -16,7 +16,7 @@
     <p>
       <ul><li><a href="/examples/http">HTTP Client</a></li></ul>
       <ul><li><a href="/examples/templates">Templates</a></li></ul>
-      <ul><li><a href="/examples/web-components">Web Components</a></li></ul>
+      <ul><li><a href="/examples/templates/web-components.html">Web Components</a></li></ul>
     </p>
   </div>
